fabric-defect-app/backend/requirements.txt
fastapi
uvicorn
python-multipart
ultralytics
opencv-python-headless
numpy
pandas
pydantic
motor
pymongo[srv]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
certifi
gunicorn
uvicorn
